move blob models
1, move project_blob into pkg
2, move artifact_blob into pkg
