--- src/common/models/project_blob.go
@@ -1,32 +1,0 @@
-// Copyright Project Harbor Authors
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//    http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package models
-
-import (
-	"time"
-)
-
-// ProjectBlob holds the relationship between manifest and blob.
-type ProjectBlob struct {
-	ID           int64     `orm:"pk;auto;column(id)" json:"id"`
-	ProjectID    int64     `orm:"column(project_id)" json:"project_id"`
-	BlobID       int64     `orm:"column(blob_id)" json:"blob_id"`
-	CreationTime time.Time `orm:"column(creation_time);auto_now_add" json:"creation_time"`
-}
-
-// TableName ...
-func (*ProjectBlob) TableName() string {
-	return "project_blob"
-}
